requirements.txt
requests==2.31.0
httpx==0.27.2
dnspython==2.6.1
beautifulsoup4==4.12.3
pandas==2.2.2
selenium==4.25.0
python-wappalyzer==0.3.1
termcolor==2.4.0
aiohttp==3.10.5
lxml==5.3.0
urllib3==2.5.0
numpy==1.26.4
tqdm==4.66.5
tldextract
flask>=3.0.0
# Advanced feature dependencies
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
shodan>=1.31.0
